@@ -27,8 +27,6 @@
     "postcss": "^8.4.12",
     "postcss-loader": "^6.2.1",
     "postcss-preset-env": "^7.4.3",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
     "react-refresh": "^0.12.0",
     "sass": "^1.49.11",
     "sass-loader": "^12.6.0",
@@ -40,6 +38,8 @@
   "dependencies": {
     "mobx": "^6.5.0",
     "mobx-react": "^7.3.0",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-router-dom": "^6.3.0"
   }
 }
