@@ -1,22 +1,22 @@
 import React from 'react'
-import { Router, Switch, Route } from 'react-router-dom'
+import { Routes, Switch, Route, BrowserRouter } from 'react-router-dom'
 // import GlobalWrapper from './core/Global';
 import { Provider } from 'mobx-react';
-import configureStore from './core/configureStore'
+// import configureStore from './core/configureStore'
+import Menu from './pages/Menu';
+import Menu2 from './pages/Menu2';
 
 const App = () => {
-  const { stores } = configureStore();
+//   const { stores } = configureStore();
 
   return (
-    <Provider {...stores}>
-          <GlobalCss />
-          <Router>
-              <GlobalWrapper>
-                <Switch>
-                  
-                </Switch>
-              </GlobalWrapper>
-          </Router>
+    <Provider>
+        <BrowserRouter>
+          <Routes>
+                <Route path={'/'} element={<Menu />} />
+                {/* <Route path={'/'} element={<Menu2 />} /> */}
+          </Routes>
+        </BrowserRouter>
     </Provider>)
 }
 
